Normalize line endings

+++ geo-project/geo-server/src/main/java/com/huanjing/geo/module/partner/entity/PartnerAccount.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,9 +14,9 @@ public class PartnerAccount {
     @TableId(type = IdType.AUTO)
     private Long id;
     private Long partnerId;
-    private Long currentBalance;
-    private Long totalRecharge;
-    private Long totalDeduction;
+    private BigDecimal currentBalance;
+    private BigDecimal totalRecharge;
+    private BigDecimal totalDeduction;
     private String currency;
     private String status;
     private LocalDateTime createdAt;
